refactor: Client가 응답으로 String을 받도록 수정한다

+++ client/src/main/kotlin/xb/dev/client/service/JoinRoomHandler.kt
@@ -25,6 +25,6 @@ internal class JoinRoomHandler(
 
         val answer=  wordMessenger.send(SupportMessage.JoinRoom(roomId, userId, token))
 
-        eventPublisher.publishEvent(JoinRoomEvent(getKey(answer)))
+        eventPublisher.publishEvent(JoinRoomEvent(getKey(answer.toInt())))
     }
 }
